storagecontrol/requirements.txt
google-cloud-storage-control>=1.11.0
